+++ src/mbedtls.c
@@ -1049,7 +1049,7 @@ cleanup:
         }                                                                   \
     } while(0)
 
-/* _libssh2_ecdsa_sign
+/* _libssh2_ecdsa_verify
  *
  * Verifies the ECDSA signature of a hashed message
  *
@@ -1331,7 +1331,7 @@ int
 _libssh2_mbedtls_ecdsa_sign(LIBSSH2_SESSION *session,
                             libssh2_ecdsa_ctx *ctx,
                             const unsigned char *hash,
-                            unsigned long hash_len,
+                            size_t hash_len,
                             unsigned char **sign,
                             size_t *sign_len)
 {
